Various fixes from the group.
------------------------------------------------------------------------
r7555 | takeshi.nakazato | 2009-04-05 20:20:46 -0600 (Sun, 05 Apr 2009) | 17 lines
Changed paths:
   M /v2/active/code/xmlcasa/scripts/task_sdfit.py
   M /v2/active/code/xmlcasa/tasks/sdfit.xml


       New Development: No
            JIRA Issue: Yes CAS-1303
      Ready to Release: Yes
     Interface Changes: Yes
What Interface Changed: Added overwrite option in the task sdfit
         Test Programs: execute sdfit with overwrite=True/False
  Put in Release Notes: Yes
             Module(s): Module Names change impacts.
           Description: 
Added 'overwrite' option for fitfile.
If overwrite=True and fitfile already exists, the task prints a message 
that informs fitfile already exists and the file is not overwritten.

	


r7556 | takeshi.nakazato | 2009-04-06 00:21:38 -0600 (Mon, 06 Apr 2009) | 20 lines
Changed paths:
   M /v2/active/code/xmlcasa/scripts/task_sdplot.py
   M /v2/active/code/xmlcasa/tasks/sdplot.xml


       New Development: No
            JIRA Issue: Yes CAS-1299
      Ready to Release: Yes
     Interface Changes: Yes
What Interface Changed: Added restfreq option in the task sdplot
         Test Programs: run sdplot with specunit='km/s' and restfreq to appropriate value
  Put in Release Notes: Yes
             Module(s): Module Names change impacts.
           Description: 
Added restfreq options in the task sdplot. This option is effective only 
when specunit is 'km/s'. If specunit is 'km/s' and restfreq is specified, 
the task changes rest frequency to the specified value before plotting data.
Allowed types for restfreq is float and string. For string, the value should 
be a numerical value or a numerical value with unit. Recognizable unit is 
'THz', 'GHz', 'MHz', 'kHz', and 'Hz'.

	


r7557 | jonas.larsen | 2009-04-06 06:00:50 -0600 (Mon, 06 Apr 2009) | 12 lines
Changed paths:
   M /v2/active/code/flagging/implement/Flagging/Flagger.cc


       New Development: No
            JIRA Issue: CAS-294
      Ready to Release: Yes
     Interface Changes: No
What Interface Changed: 
         Test Programs: flagdata_test.py
  Put in Release Notes: No
             Module(s):
           Description: Fixed the counting of the amount of data flags
                        (by looping over polarization setups too)

r7558 | jonas.larsen | 2009-04-06 06:22:29 -0600 (Mon, 06 Apr 2009) | 11 lines
Changed paths:
   M /v2/active/code/xmlcasa/implement/flagging/flagger_cmpt.cc
   M /v2/active/code/xmlcasa/scripts/task_flagmanager.py


       New Development: No
            JIRA Issue: CAS-377
      Ready to Release: Yes
     Interface Changes: No
What Interface Changed: 
         Test Programs: No
  Put in Release Notes: No
             Module(s):
           Description: Send flagmanager(mode='list') output to logger, not to standard output.

r7559 | wes.young | 2009-04-06 10:20:59 -0600 (Mon, 06 Apr 2009) | 14 lines
Changed paths:
   M /v2/active/code/msvis/apps/casasplit/makefile


       New Development: No
            JIRA Issue: No
      Ready to Release: Yes
     Interface Changes: No
What Interface Changed: 
         Test Programs: 
  Put in Release Notes: No
             Module(s): msvis
           Description: 
Removed dependency on flagging library.
	


r7562 | shannon.jaeger | 2009-04-06 16:08:39 -0600 (Mon, 06 Apr 2009) | 22 lines
Changed paths:
   M /v2/active/code/xmlcasa/implement/images/coordsys_cmpt.cc


       New Development: No
            JIRA Issue: Yes  - CAS-1093
      Ready to Release: Yes
     Interface Changes: No
What Interface Changed: N/A
         Test Programs: None - maybe Raymond's image analysis test
                        xmlcasa/implement/images/test/*.py
  Put in Release Notes: No
             Module(s): coordinate system tool
           Description: Changed setlneartransform to accept
			both double and integer vectors for
			input.

			May need to add support for complex numbered
			vectors at some point.

			There was no type check being done but a double
			vector was expected, hence errors occured otherwise.
			


r7563 | laura.glendenning | 2009-04-06 16:52:40 -0600 (Mon, 06 Apr 2009) | 13 lines
Changed paths:
   M /v2/active/code/plotms/implement/GuiTabs/PlotMSPlotTab.ui


       New Development: No
            JIRA Issue: No
      Ready to Release: Yes
     Interface Changes: No
What Interface Changed: N/A
         Test Programs: N/A
  Put in Release Notes: No
             Module(s): plotms
           Description: Moved baseline averaging widget to not be dependent on
                        time averaging widget.


r7564 | wes.young | 2009-04-06 17:19:54 -0600 (Mon, 06 Apr 2009) | 14 lines
Changed paths:
   M /v2/active/code/xmlcasa/scripts/casapy.py


       New Development: No
            JIRA Issue: CAS-556
      Ready to Release: Yes
     Interface Changes: No
What Interface Changed: 
         Test Programs: None
  Put in Release Notes: No
             Module(s): casapy
           Description: 
Added some explanatory text to the casalogger definition.




git-svn-id: https://svn.cv.nrao.edu/svn/casa/v2/stable@7566 ac9c0da5-131f-0410-9009-a9c195f183e6

+++ code/xmlcasa/scripts/task_sdfit.py
@@ -4,7 +4,7 @@ from taskinit import *
 import asap as sd
 import pylab as pl
 
-def sdfit(sdfile, fluxunit, telescopeparm, specunit, frame, doppler, scanlist, field, iflist, pollist, fitmode, maskline, invertmask, nfit, thresh, min_nchan, avg_limit, box_size, edge, fitfile, plotlevel):
+def sdfit(sdfile, fluxunit, telescopeparm, specunit, frame, doppler, scanlist, field, iflist, pollist, fitmode, maskline, invertmask, nfit, thresh, min_nchan, avg_limit, box_size, edge, fitfile, overwrite, plotlevel):
 
 
         casalog.origin('sdfit')
@@ -492,8 +492,8 @@ def sdfit(sdfile, fluxunit, telescopeparm, specunit, frame, doppler, scanlist, f
             # Store fit
             if ( fitfile != '' ):
                     #f.store_fit(fitfile)
-                    print fitparams
-                    store_fit("gauss", fitfile, fitparams, s)
+                    #print fitparams
+                    store_fit("gauss", fitfile, fitparams, s, overwrite)
 
                 
             # Final clean up
@@ -513,11 +513,15 @@ def sdfit(sdfile, fluxunit, telescopeparm, specunit, frame, doppler, scanlist, f
                 
 
 ### store_fit
-def store_fit(func, fitfile, value, scan):
+def store_fit(func, fitfile, value, scan, overwrite):
         outname=os.path.expandvars(fitfile)
         outname=os.path.expanduser(outname)
         if os.path.exists(outname):
-                os.system('rm '+outname)
+                if overwrite:
+                        os.system('rm '+outname)
+                else:
+                        print fitfile+' already exists.'
+                        return
                 
         outfile=file(outname, 'w')
